+++ NEWS.md
@@ -19,8 +19,11 @@ If you defined your own manifolds and used traits, please check the documentatio
 ### Changed
 
 * refactor the trait system to no longer use a list of traits but single traits separately for the metric and the embedding specification
-* Introduce an `is_embedded` trait to check if a manifold is embedded.
 * Switch to using [Runic.jl](https://github.com/fredrikekre/Runic.jl) as code formatter
+
+### Removed
+
+* `ODEExponentialRetraction` was removed in favor of `solve_chart_exp_ode` implemented in `Manifolds.jl`.
 
 ## [1.2.0] 08/05/2025